brod vmaster brod_cg_commits

This is a utility module to help force commit offsets to kafka.

Types

group_id()

group_id() :: brod:group_id().

group_input()

group_input() :: [{prop_key(), prop_val()}].

member_id()

member_id() :: brod:group_member_id().

offset()

offset() :: brod:offset().

offsets()

offsets() :: latest | earliest | [{partition(), offset()}].

partition()

partition() :: brod:partition().

prop_key()

prop_key() :: id | topic | retention | protocol | offsets.

prop_val()

prop_val() ::
    group_id() |
    topic() |
    retention() |
    offsets() |
    brod_group_coordinator:protocol_name().

retention()

retention() :: integer().

-1 to use whatever configured in kafka

topic()

topic() :: brod:topic().

Functions

assign_partitions/3
(since )

assign_partitions(pid(),
                  [brod:group_member()],
                  [{brod:topic(), brod:partition()}]) ->
                     [{member_id(), [brod:partition_assignment()]}].

This function is called only when partition_assignment_strategy is set for callback_implemented in group config.

assignments_received/4
(since )

assignments_received(pid(),
                     member_id(),
                     integer(),
                     brod:received_assignments()) ->
                        ok.

Called by group coordinator when there is new assignemnt received.

assignments_revoked/1
(since )

assignments_revoked(pid()) -> ok.

Called by group coordinator before re-joinning the consumer group.

code_change/3
(since )

Callback implementation for c::gen_server.code_change/3.

get_committed_offsets/2
(since )

get_committed_offsets(pid(), [{brod:topic(), brod:partition()}]) ->
                         {ok,
                          [{{brod:topic(), brod:partition()},
                            brod:offset()}]}.

Called by group coordinator when initializing the assignments for subscriber. NOTE: this function is called only when it is DISABLED to commit offsets to kafka. i.e. offset_commit_policy is set to consumer_managed

handle_call/3
(since )

Callback implementation for c::gen_server.handle_call/3.

handle_cast/2
(since )

Callback implementation for c::gen_server.handle_cast/2.

handle_info/2
(since )

Callback implementation for c::gen_server.handle_info/2.

init/1
(since )

Callback implementation for c::gen_server.init/1.

run/2
(since )

Force commit offsets.

start_link/2
(since )

start_link(brod:client(), group_input()) ->
              {ok, pid()} | {error, any()}.

Start (link) a group member. The member will try to join the consumer group and get assignments for the given topic-partitions, then commit given offsets to kafka. In case not all given partitions are assigned to it, it will terminate with an exit exception

stop/1
(since )

stop(pid()) -> ok.

Stop the process.

sync/1
(since )

sync(pid()) -> ok.

Make a call to the resetter process, the call will be blocked until offsets are committed.

terminate/2
(since )

Callback implementation for c::gen_server.terminate/2.